fix: Processing changes
Renamed stream_transcript_job.rb to incident_replay_job.rb to replace the one we deleted.

Updated IncidentReplayJob to use the database as the source of truth. Prevented the double & conflicting sources of truth that was there before

Modified IncidentsController to trigger IncidentReplayJob instead of the old job and improved the transcript processing logic as a result.